Edit search action (task #2802)
Search action requires two parameters, the id of the edited search
and the id of the pre-saved search. We fetch the pre-saved search
entity by its id, we merge it with the request data (which might
include a new name for the saved search) and then we patch the
existing saved search entity with the newly merged array.